Rename capabilities folder
